# internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NovelForge Local API",
    "version": "0.1.0-alpha.3",
    "description": "Project lifecycle, formal local Web workspace, durable events and engine adapter."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:48090"
    }
  ],
  "paths": {
...
    "/api/events": {
      "get": {
        "operationId": "streamEvents",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Bad"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "parameters": [
          {
            "name": "Last-Event-ID",
            "in": "header",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "last_event_id",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "project",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ]
      }
    },
...
    "/api/truth/events": {
      "get": {
        "operationId": "listTruthEvents",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "after_sequence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "through_chapter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "subject_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "pattern": "^[A-Za-z][A-Za-z0-9_.:-]{0,127}$"
            }
          },
          {
            "name": "subject_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "predicate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "pattern": "^[A-Za-z][A-Za-z0-9_.:-]{0,127}$"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Truth event page in stable sequence order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruthEventPage"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "appendTruthEvent",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TruthAppendRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Truth event appended atomically",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruthAppendResult"
                }
              }
            }
          },
          "200": {
            "description": "Exact response replayed from the persistent idempotency ledger",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruthAppendResult"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/truth/state": {
      "get": {
        "operationId": "queryTruthState",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "chapter",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "subject_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "pattern": "^[A-Za-z][A-Za-z0-9_.:-]{0,127}$"
            }
          },
          {
            "name": "subject_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "predicate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "pattern": "^[A-Za-z][A-Za-z0-9_.:-]{0,127}$"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chapter-bounded Truth projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruthStatePage"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/truth/state:batch": {
      "post": {
        "operationId": "batchQueryTruthState",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TruthBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch Chapter-N Truth projections in request order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruthBatchResponse"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/truth/conflicts": {
      "get": {
        "operationId": "listTruthConflicts",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "chapter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "subject_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "pattern": "^[A-Za-z][A-Za-z0-9_.:-]{0,127}$"
            }
          },
          {
            "name": "subject_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "predicate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "pattern": "^[A-Za-z][A-Za-z0-9_.:-]{0,127}$"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 10,
              "enum": [
                "unresolved",
                "resolved"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stable, paginated Truth conflict history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruthConflictPage"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/truth/rebuild": {
      "post": {
        "operationId": "rebuildTruthProjection",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TruthRebuildRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Projection rebuilt from the requested chapter boundary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruthRebuildResult"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/truth/verify": {
      "get": {
        "operationId": "verifyTruthProjection",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event and projection integrity verification",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TruthVerifyResult"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
        }
      },
...
    },
    "responses": {
      "Bad": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "Internal": {
        "description": "Internal error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "details",
          "retryable",
          "trace_id"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          },
          "retryable": {
            "type": "boolean"
          },
          "trace_id": {
            "type": "string"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
...
      "TruthAppendRequest": {
        "type": "object",
        "required": [
          "project_id",
          "event"
        ],
        "properties": {
          "project_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "event": {
            "$ref": "#/components/schemas/TruthEventInput"
          }
        },
        "additionalProperties": false
      },
      "TruthEvent": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TruthEventInput"
          },
          {
            "type": "object",
            "required": [
              "sequence",
              "id",
              "created_at",
              "checksum"
            ],
            "properties": {
              "sequence": {
                "type": "integer"
              },
              "id": {
                "type": "string"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "checksum": {
                "type": "string"
              }
            }
          }
        ]
      },
      "TruthFact": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TruthEvent"
          },
          {
            "type": "object",
            "required": [
              "effective_from_chapter",
              "conflicted"
            ],
            "properties": {
              "effective_from_chapter": {
                "type": "integer"
              },
              "effective_to_chapter": {
                "type": [
                  "integer",
                  "null"
                ]
              },
              "superseded_by_event_id": {
                "type": "string"
              },
              "conflicted": {
                "type": "boolean"
              }
            }
          }
        ]
      },
...
      "TruthRebuildRequest": {
        "type": "object",
        "required": [
          "project_id",
          "from_chapter"
        ],
        "properties": {
          "project_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "from_chapter": {
            "type": "integer",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "TruthStateQuery": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "chapter"
        ],
        "properties": {
          "chapter": {
            "type": "integer",
            "minimum": 0
          },
          "subject_type": {
            "type": "string",
            "maxLength": 128,
            "pattern": "^[A-Za-z][A-Za-z0-9_.:-]{0,127}$"
          },
          "subject_id": {
            "type": "string",
            "maxLength": 200
          },
          "predicate": {
            "type": "string",
            "maxLength": 128,
            "pattern": "^[A-Za-z][A-Za-z0-9_.:-]{0,127}$"
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 500,
            "default": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0,
            "default": 0
          }
        }
      },
      "TruthBatchRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "project_id",
          "queries"
        ],
        "properties": {
          "project_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "queries": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/TruthStateQuery"
            }
          }
        }
      },
      "TruthAppendResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "event",
          "replayed"
        ],
        "properties": {
          "event": {
            "$ref": "#/components/schemas/TruthEvent"
          },
          "replayed": {
            "type": "boolean"
          }
        }
      },
      "TruthEventPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "events",
          "after_sequence",
          "limit"
        ],
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TruthEvent"
            }
          },
          "after_sequence": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 500
          },
          "next_sequence": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "TruthStatePage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "facts",
          "conflicts",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "facts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TruthFact"
            }
          },
          "conflicts": {
            "type": "integer",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 500
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          },
          "next_offset": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "TruthBatchResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "results"
        ],
        "properties": {
          "results": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/TruthStatePage"
            }
          }
        }
      },
      "TruthConflictPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "conflicts",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "conflicts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TruthConflict"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 500
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          },
          "next_offset": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "TruthRebuildResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "from_chapter",
          "events_replayed",
          "facts_projected",
          "conflicts_projected",
          "projection_digest"
        ],
        "properties": {
          "from_chapter": {
            "type": "integer",
            "minimum": 0
          },
          "events_replayed": {
            "type": "integer",
            "minimum": 0
          },
          "facts_projected": {
            "type": "integer",
            "minimum": 0
          },
          "conflicts_projected": {
            "type": "integer",
            "minimum": 0
          },
          "projection_digest": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          }
        }
      },
      "TruthVerifyResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "events",
          "facts",
          "conflicts",
          "projection_digest",
          "valid",
          "violations"
        ],
        "properties": {
          "events": {
            "type": "integer",
            "minimum": 0
          },
          "facts": {
            "type": "integer",
            "minimum": 0
          },
          "conflicts": {
            "type": "integer",
            "minimum": 0
          },
          "projection_digest": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "valid": {
            "type": "boolean"
          },
          "violations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}